macOS: extract SurfaceConfiguration into its own file
SurfaceConfiguration is a model type, not a view, and it was sitting in the
middle of SurfaceView.swift between two overlay views. Move it to
SurfaceConfiguration.swift alongside the rest of the surface sources.

It only needs AppKit, GhosttyKit and System, so the new file imports those
rather than SwiftUI.

Pure code motion.
