+++ macos/Sources/Ghostty/Surface View/SurfaceConfiguration.swift
@@ -1,0 +1,119 @@
+import AppKit
+import GhosttyKit
+import System
+
+extension Ghostty {
+    /// The configuration for a surface. For any configuration not set, defaults will be chosen from
+    /// libghostty, usually from the Ghostty configuration.
+    struct SurfaceConfiguration {
+        /// Explicit font size to use in points
+        var fontSize: Float32?
+
+        /// Explicit working directory. This is normalized on assignment to
+        /// remove any redundant and trailing path separators.
+        var workingDirectory: String? {
+            get { normalizedWorkingDirectory }
+            set { normalizedWorkingDirectory = newValue.map { FilePath($0).string } }
+        }
+        private var normalizedWorkingDirectory: String?
+
+        /// Explicit command to set
+        var command: String?
+
+        /// Environment variables to set for the terminal
+        var environmentVariables: [String: String] = [:]
+
+        /// Extra input to send as stdin
+        var initialInput: String?
+
+        /// Wait after the command
+        var waitAfterCommand: Bool = false
+
+        /// Context for surface creation
+        var context: ghostty_surface_context_e = GHOSTTY_SURFACE_CONTEXT_WINDOW
+
+        init() {}
+
+        init(from config: ghostty_surface_config_s) {
+            self.fontSize = config.font_size
+            if let workingDirectory = config.working_directory {
+                self.workingDirectory = String.init(cString: workingDirectory, encoding: .utf8)
+            }
+            if let command = config.command {
+                self.command = String.init(cString: command, encoding: .utf8)
+            }
+
+            // Convert the C env vars to Swift dictionary
+            if config.env_var_count > 0, let envVars = config.env_vars {
+                for i in 0..<config.env_var_count {
+                    let envVar = envVars[i]
+                    if let key = String(cString: envVar.key, encoding: .utf8),
+                       let value = String(cString: envVar.value, encoding: .utf8) {
+                        self.environmentVariables[key] = value
+                    }
+                }
+            }
+            self.context = config.context
+        }
+
+        /// Provides a C-compatible ghostty configuration within a closure. The configuration
+        /// and all its string pointers are only valid within the closure.
+        func withCValue<T>(view: SurfaceView, _ body: (inout ghostty_surface_config_s) throws -> T) rethrows -> T {
+            var config = ghostty_surface_config_new()
+            config.userdata = Unmanaged.passUnretained(view).toOpaque()
+            config.platform_tag = GHOSTTY_PLATFORM_MACOS
+            config.platform = ghostty_platform_u(macos: ghostty_platform_macos_s(
+                nsview: Unmanaged.passUnretained(view).toOpaque()
+            ))
+            config.scale_factor = NSScreen.main!.backingScaleFactor
+
+            // Zero is our default value that means to inherit the font size.
+            config.font_size = fontSize ?? 0
+
+            // Set wait after command
+            config.wait_after_command = waitAfterCommand
+
+            // Set context
+            config.context = context
+
+            // Use withCString to ensure strings remain valid for the duration of the closure
+            return try workingDirectory.withCString { cWorkingDir in
+                config.working_directory = cWorkingDir
+
+                return try command.withCString { cCommand in
+                    config.command = cCommand
+
+                    return try initialInput.withCString { cInput in
+                        config.initial_input = cInput
+
+                        // Convert dictionary to arrays for easier processing
+                        let keys = Array(environmentVariables.keys)
+                        let values = Array(environmentVariables.values)
+
+                        // Create C strings for all keys and values
+                        return try keys.withCStrings { keyCStrings in
+                            return try values.withCStrings { valueCStrings in
+                                // Create array of ghostty_env_var_s
+                                var envVars = [ghostty_env_var_s]()
+                                envVars.reserveCapacity(environmentVariables.count)
+                                for i in 0..<environmentVariables.count {
+                                    envVars.append(ghostty_env_var_s(
+                                        key: keyCStrings[i],
+                                        value: valueCStrings[i]
+                                    ))
+                                }
+
+                                return try envVars.withUnsafeMutableBufferPointer { buffer in
+                                    config.env_vars = buffer.baseAddress
+                                    config.env_var_count = environmentVariables.count
+                                    return try body(&config)
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
